pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.1"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
nanoid = "^2.0.0"
Jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
google-cloud-datastore = {version = "^2.5.1", optional = true}
redis = {version = "^4.2.2", optional = true}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pytest = "^5.2"
mypy = "^0.950"

[tool.poetry.extras]
redis = ["redis"]
google-cloud-datastore = ["google-cloud-datastore"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
